Serial/TCP/UDP ports: mark messages as forwarded
When we forward a MAVLink message we should not change the sequence
number. With this change we are doing that and avoid confusing PX4
which has to assume messages were dropped if the sequence number is
not correctly increasing.

By marking the messages as forwarded we prevent the sequence number from
being changed in the first place.